# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go Gin personal blog backend API.",
        "title": "MyBlog API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/admin/articles": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员分页获取后台文章列表，支持状态、删除状态、分类和关键词筛选。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin/article"
                ],
                "summary": "获取后台文章列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码，默认 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认 10",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "文章状态，-1 全部，0 草稿，1 发布，2 下架",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "删除状态，-1 全部，0 未删除，1 已删除",
                        "name": "is_deleted",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "分类 ID，0 表示全部",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "关键词",
                        "name": "keyword",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "示例：{\\\"code\\\":200,\\\"data\\\":{\\\"list\\\":[],\\\"total\\\":0,\\\"page\\\":1,\\\"size\\\":10},\\\"msg\\\":\\\"获取后台文章列表成功\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "示例：{\\\"code\\\":400,\\\"msg\\\":\\\"文章状态参数错误\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "示例：{\\\"code\\\":500,\\\"msg\\\":\\\"获取后台文章列表失败\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/api/articles/search": {
            "get": {
                "description": "根据关键词分页搜索已发布且未删除的文章。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article"
                ],
                "summary": "搜索文章",
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码，默认 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认 10",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "示例：{\\\"code\\\":200,\\\"data\\\":{\\\"list\\\":[],\\\"total\\\":0,\\\"page\\\":1,\\\"size\\\":10},\\\"msg\\\":\\\"搜索成功\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "示例：{\\\"code\\\":500,\\\"msg\\\":\\\"搜索文章失败\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/articles/timeline": {
            "get": {
                "description": "分页获取已发布且未删除的文章时间轴列表。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article"
                ],
                "summary": "获取文章时间轴",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码，默认 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认 10",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "示例：{\\\"code\\\":200,\\\"data\\\":{\\\"list\\\":[],\\\"total\\\":0},\\\"msg\\\":\\\"获取文章时间轴成功\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "示例：{\\\"code\\\":500,\\\"msg\\\":\\\"获取文章时间轴失败\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/api/categories/{id}/articles": {
            "get": {
                "description": "根据分类 ID 分页获取已发布且未删除的文章列表。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article/category"
                ],
                "summary": "按分类获取文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分类 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码，默认 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认 10",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "示例：{\\\"code\\\":200,\\\"data\\\":{\\\"list\\\":[],\\\"total\\\":0},\\\"msg\\\":\\\"获取文章列表成功\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "示例：{\\\"code\\\":400,\\\"msg\\\":\\\"分类ID不合法\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "示例：{\\\"code\\\":500,\\\"msg\\\":\\\"获取文章列表失败\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/api/tags/{id}/articles": {
            "get": {
                "description": "根据标签 ID 分页获取已发布且未删除的文章列表。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article/tag"
                ],
                "summary": "按标签获取文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "标签 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码，默认 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认 10",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "示例：{\\\"code\\\":200,\\\"data\\\":{\\\"list\\\":[],\\\"total\\\":0},\\\"msg\\\":\\\"获取标签文章列表成功\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "示例：{\\\"code\\\":400,\\\"msg\\\":\\\"标签ID不合法\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "示例：{\\\"code\\\":500,\\\"msg\\\":\\\"获取标签文章列表失败\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}